Data.Digest.Keccak

Description

Crypto algorithms of the Keccak family (SHA3/SHAKE), with support for GHCJS.

Documentation

sha3_256 :: Int -> ByteString -> ByteString

Return the specified number of bytes of the SHA-3 hash of the provided byte string

shake_128 :: Int -> ByteString -> ByteString

Return the specified number of bytes of the SHAKE-128 hash of the provided byte string >>> shake_128 8 B.empty == B.pack [127, 156, 43, 164, 232, 143, 130, 125] True

>>> shake_128 32 (B.pack [1..10]) == B.pack [142,56,168,122,207,188,35,211,233,209,95,158,63,91,102,156,114,204,22,38,177,105,130,116,173,114,190,153,159,101,10,150]
True

>>> let i = B.pack [1..10] in shake_128 4 i == B.take 4 (shake_128 32 i)
True