Data.These.Extra

Documentation

data These a b

aThese :: Maybe a -> Maybe b -> These a b

A data type that represents either a value of type this, or a value of type that, or both. data These this that = This this | That that | These this that deriving (Eq,Ord, Read, Show)

Constructor function, from Maybes

both :: Semigroup a => (t1 -> a) -> (t2 -> a) -> These t1 t2 -> a

hasThis :: These a b -> Bool

hasThat :: These a1 a -> Bool

fromThis :: These c b -> c

fromThat :: These a1 c -> c