ZM

Documentation

Check the tutorial and github repo.

class Generic a where

Representable types of kind *. This class is derivable in GHC with the DeriveGeneric flag on.

A Generic instance must satisfy the following laws:

from . to ≡ id
to . from ≡ id

Associated Types

type Rep a :: Type -> Type

Generic representation type

Methods

from :: a -> Rep a x

Convert from the datatype to its representation

to :: Rep a x -> a

Convert from the representation to the datatype

class Generic1 (f :: k -> Type) where

Representable types of kind * -> * (or kind k -> *, when PolyKinds is enabled). This class is derivable in GHC with the DeriveGeneric flag on.

A Generic1 instance must satisfy the following laws:

from1 . to1 ≡ id
to1 . from1 ≡ id

Associated Types

type Rep1 (f :: k -> Type) :: k -> Type

Generic representation type

Methods

from1 :: forall (a :: k). f a -> Rep1 f a

Convert from the datatype to its representation

to1 :: forall (a :: k). Rep1 f a -> f a

Convert from the representation to the datatype

class Datatype (d :: k) where

Class for datatypes that represent datatypes

Minimal complete definition

datatypeName, moduleName, packageName

Methods

datatypeName :: forall k1 t (f :: k1 -> Type) (a :: k1). t d f a -> [Char]

The name of the datatype (unqualified)

moduleName :: forall k1 t (f :: k1 -> Type) (a :: k1). t d f a -> [Char]

The fully-qualified name of the module where the type is declared

packageName :: forall k1 t (f :: k1 -> Type) (a :: k1). t d f a -> [Char]

The package name of the module where the type is declared

Since: base-4.9.0.0

isNewtype :: forall k1 t (f :: k1 -> Type) (a :: k1). t d f a -> Bool

Marks if the datatype is actually a newtype

Since: base-4.7.0.0

class Constructor (c :: k) where

Class for datatypes that represent data constructors

Minimal complete definition

conName

Methods

conName :: forall k1 t (f :: k1 -> Type) (a :: k1). t c f a -> [Char]

The name of the constructor

conFixity :: forall k1 t (f :: k1 -> Type) (a :: k1). t c f a -> Fixity

The fixity of the constructor

conIsRecord :: forall k1 t (f :: k1 -> Type) (a :: k1). t c f a -> Bool

Marks if this constructor is a record

class Selector (s :: k) where

Class for datatypes that represent records

Methods

selName :: forall k1 t (f :: k1 -> Type) (a :: k1). t s f a -> [Char]

The name of the selector

selSourceUnpackedness :: forall k1 t (f :: k1 -> Type) (a :: k1). t s f a -> SourceUnpackedness

The selector's unpackedness annotation (if any)

Since: base-4.9.0.0

selSourceStrictness :: forall k1 t (f :: k1 -> Type) (a :: k1). t s f a -> SourceStrictness

The selector's strictness annotation (if any)

Since: base-4.9.0.0

selDecidedStrictness :: forall k1 t (f :: k1 -> Type) (a :: k1). t s f a -> DecidedStrictness

The strictness that the compiler inferred for the selector

Since: base-4.9.0.0

data V1 (p :: k)

Void: used for datatypes without constructors

data U1 (p :: k)

Unit: used for constructors without arguments

Constructors

U1

newtype Par1 p

Used for marking occurrences of the parameter

Constructors

Par1

Fields

unPar1 :: p

newtype Rec1 (f :: k -> Type) (p :: k)

Recursive calls of kind * -> * (or kind k -> *, when PolyKinds is enabled)

Constructors

Rec1

Fields

unRec1 :: f p

newtype K1 i c (p :: k)

Constants, additional parameters and recursion of kind *

Constructors

K1

Fields

unK1 :: c

newtype M1 i (c :: Meta) (f :: k -> Type) (p :: k)

Meta-information (constructor names, etc.)

Constructors

M1

Fields

unM1 :: f p

data ((f :: k -> Type) :+: (g :: k -> Type)) (p :: k) infixr 5

Sums: encode choice between constructors

Constructors

L1 (f p)

R1 (g p)

data ((f :: k -> Type) :*: (g :: k -> Type)) (p :: k) infixr 6

Products: encode multiple arguments to constructors

Constructors

(f p) :*: (g p) infixr 6

newtype ((f :: k2 -> Type) :.: (g :: k1 -> k2)) (p :: k1) infixr 7

Composition of functors

Constructors

Comp1

Fields

unComp1 :: f (g p)

data R

Tag for K1: recursion (of kind Type)

data D

Tag for M1: datatype

data C

Tag for M1: constructor

data S

Tag for M1: record selector

type Rec0 = K1 R :: Type -> k -> Type

Type synonym for encoding recursion (of kind Type)

type D1 = M1 D :: Meta -> (k -> Type) -> k -> Type

Type synonym for encoding meta-information for datatypes

type C1 = M1 C :: Meta -> (k -> Type) -> k -> Type

Type synonym for encoding meta-information for constructors

type S1 = M1 S :: Meta -> (k -> Type) -> k -> Type

Type synonym for encoding meta-information for record selectors

type family Rep a :: Type -> Type

Generic representation type

type family Rep1 (f :: k -> Type) :: k -> Type

Generic representation type

data family URec a (p :: k)

Constants of unlifted kinds

Since: base-4.9.0.0

type UAddr = URec (Ptr ()) :: k -> Type

Type synonym for URec Addr#

Since: base-4.9.0.0

type UChar = URec Char :: k -> Type

Type synonym for URec Char#

Since: base-4.9.0.0

type UDouble = URec Double :: k -> Type

Type synonym for URec Double#

Since: base-4.9.0.0

type UFloat = URec Float :: k -> Type

Type synonym for URec Float#

Since: base-4.9.0.0

type UInt = URec Int :: k -> Type

Type synonym for URec Int#

Since: base-4.9.0.0

type UWord = URec Word :: k -> Type

Type synonym for URec Word#

Since: base-4.9.0.0

prec :: Fixity -> Int

Get the precedence of a fixity value.

data Fixity

Datatype to represent the fixity of a constructor. An infix | declaration directly corresponds to an application of Infix.

Constructors

Prefix

Infix Associativity Int

data FixityI

This variant of Fixity appears at the type level.

Since: base-4.9.0.0

Constructors

PrefixI

InfixI Associativity Nat

data Associativity

Datatype to represent the associativity of a constructor

Constructors

LeftAssociative

RightAssociative

NotAssociative

data SourceUnpackedness

The unpackedness of a field as the user wrote it in the source code. For example, in the following data type:

data E = ExampleConstructor     Int
           {-# NOUNPACK #-} Int
           {-#   UNPACK #-} Int

The fields of ExampleConstructor have NoSourceUnpackedness, SourceNoUnpack, and SourceUnpack, respectively.

Since: base-4.9.0.0

Constructors

NoSourceUnpackedness

SourceNoUnpack

SourceUnpack

data SourceStrictness

The strictness of a field as the user wrote it in the source code. For example, in the following data type:

data E = ExampleConstructor Int ~Int !Int

The fields of ExampleConstructor have NoSourceStrictness, SourceLazy, and SourceStrict, respectively.

Since: base-4.9.0.0

Constructors

NoSourceStrictness

SourceLazy

SourceStrict

data DecidedStrictness

The strictness that GHC infers for a field during compilation. Whereas there are nine different combinations of SourceUnpackedness and SourceStrictness, the strictness that GHC decides will ultimately be one of lazy, strict, or unpacked. What GHC decides is affected both by what the user writes in the source code and by GHC flags. As an example, consider this data type:

data E = ExampleConstructor {-# UNPACK #-} !Int !Int Int

If compiled without optimization or other language extensions, then the fields of ExampleConstructor will have DecidedStrict, DecidedStrict, and DecidedLazy, respectively.

If compiled with -XStrictData enabled, then the fields will have DecidedStrict, DecidedStrict, and DecidedStrict, respectively.

If compiled with -O2 enabled, then the fields will have DecidedUnpack, DecidedStrict, and DecidedLazy, respectively.

Since: base-4.9.0.0

Constructors

DecidedLazy

DecidedStrict

DecidedUnpack

data Meta

Datatype to represent metadata associated with a datatype (MetaData), constructor (MetaCons), or field selector (MetaSel).

In MetaData n m p nt, n is the datatype's name, m is the module in which the datatype is defined, p is the package in which the datatype is defined, and nt is 'True if the datatype is a newtype.

In MetaCons n f s, n is the constructor's name, f is its fixity, and s is 'True if the constructor contains record selectors.

In MetaSel mn su ss ds, if the field uses record syntax, then mn is Just the record name. Otherwise, mn is Nothing. su and ss are the field's unpackedness and strictness annotations, and ds is the strictness that GHC infers for the field.

Since: base-4.9.0.0

Constructors

MetaData Symbol Symbol Symbol Bool

MetaCons Symbol FixityI Bool

MetaSel (Maybe Symbol) SourceUnpackedness SourceStrictness DecidedStrictness

data Proxy (t :: k)

Proxy is a type that holds no data, but has a phantom parameter of arbitrary type (or even kind). Its use is to provide type information, even though there is no value available of that type (or it may be too costly to create one).

Historically, Proxy :: Proxy a is a safer alternative to the undefined :: a idiom.

>>> Proxy :: Proxy (Void, Int -> Int)
Proxy

Proxy can even hold types of higher kinds,

>>> Proxy :: Proxy Either
Proxy

>>> Proxy :: Proxy Functor
Proxy

>>> Proxy :: Proxy complicatedStructure
Proxy

Constructors

Proxy

prettyShow :: Pretty a => a -> String

Pretty print a value with the prettyNormal level.

class Pretty a where

Pretty printing class. The precedence level is used in a similar way as in the Show class. Minimal complete definition is either pPrintPrec or pPrint.

Minimal complete definition

pPrintPrec | pPrint

Methods

pPrintPrec :: PrettyLevel -> Rational -> a -> Doc

pPrint :: a -> Doc

pPrintList :: PrettyLevel -> [a] -> Doc

convert :: Convertible a b => a -> b

typeModel :: AsType (Ana a) => Proxy a -> HTypeModel

useCT :: Typeable a => Maybe (ConTree String HTypeRef) -> proxy a -> State Env (Type (TypeRef QualName))

dottedP :: [String] -> Doc

prettyADT :: (Pretty name, Pretty consName, Pretty ref) => String -> Char -> ADT name consName ref -> Doc

spacedP :: Pretty a => [a] -> Doc

varC :: Integral a => a -> Char

varP :: Integral n => n -> Doc

vspacedP :: Pretty a => [a] -> Doc

adtNamesMap :: (adtName1 -> adtName2) -> (consName1 -> consName2) -> ADT adtName1 consName1 ref -> ADT adtName2 consName2 ref

conTreeNameFold :: Monoid a => (name -> a) -> ConTree name t -> a

conTreeNameMap :: (name -> name2) -> ConTree name t -> ConTree name2 t

conTreeTypeFoldMap :: Monoid a => (Type t -> a) -> ConTree name t -> a

conTreeTypeList :: ConTree name t -> [Type t]

conTreeTypeMap :: (Type t -> Type ref) -> ConTree name t -> ConTree name ref

constructorInfo :: Eq consName => consName -> ConTree consName t -> Maybe ([Bool], [Type t])

constructors :: ConTree name ref -> [(name, Fields name ref)]

contree :: [(name, Fields name ref)] -> Maybe (ConTree name ref)

fieldsNames :: Either t [(a, t1)] -> [t1]

fieldsTypes :: Either [b] [(a, b)] -> [b]

getHRef :: TypeRef a -> Maybe a

nestedTypeNs :: TypeN t -> [TypeN t]

qualName :: QualName -> String

solve :: (Ord k, Show k) => k -> Map k a -> a

solveAll :: (Functor f, Show k, Ord k) => Map k b -> f k -> f b

typeA :: TypeN ref -> Type ref

typeADTs :: TypeModel adtName consName inRef k -> [ADT adtName consName inRef]

typeN :: Type r -> TypeN r

unVar :: TypeRef t -> t

convertOrError :: Convertible a c => a -> Either String c

convertResultToError :: Bifunctor p => p ConvertError c -> p String c

convertResultToErrors :: Bifunctor p => p ConvertError c -> p [String] c

dotted :: [String] -> String

errorToConvertResult :: (Typeable b, Typeable a, Show a) => (a -> Either Error b) -> a -> ConvertResult b

errorsToConvertResult :: (Typeable b, Typeable t, Show t) => (t -> Either Errors b) -> t -> ConvertResult b

errsInContext :: (Convertible ctx String, Bifunctor p) => ctx -> p [String] c -> p [String] c

inContext :: Convertible ctx String => ctx -> [String] -> [String]

noErrors :: Errors -> Bool

toErrors :: Bifunctor p => p a c -> p [a] c

data ConvertError

Constructors

ConvertError

Fields

convSourceValue :: String

convSourceType :: String

convDestType :: String

convErrorMessage :: String

type ConvertResult a = Either ConvertError a

class Convertible a b where

Methods

safeConvert :: a -> ConvertResult b

class AsType a where

Methods

asType :: a -> State Env HType

class (Typeable a, AsType (Ana a)) => Model a where

Minimal complete definition

Nothing

Methods

envType :: Proxy a -> State Env HType

newtype CompactPretty a

Constructors

CompactPretty a

data ADT name consName ref

Constructors

ADT

Fields

declName :: name

declNumParameters :: Word8

declCons :: Maybe (ConTree consName ref)

data ConTree name ref

Constructors

Con

Fields

constrName :: name

constrFields :: Fields name ref

ConTree (ConTree name ref) (ConTree name ref)

type Fields name ref = Either [Type ref] [(name, Type ref)]

type HADT = ADT String String HTypeRef

type HType = Type HTypeRef

type HTypeEnv = TypeEnv String String (TypeRef QualName) QualName

type HTypeModel = TypeModel String String (TypeRef QualName) QualName

type HTypeRef = TypeRef QualName

data QualName

Constructors

QualName

Fields

pkgName :: String

mdlName :: String

locName :: String

data Type ref

Constructors

TypeCon ref

TypeApp (Type ref) (Type ref)

type TypeEnv adtName consName inRef exRef = Map exRef (ADT adtName consName inRef)

data TypeModel adtName consName inRef exRef

Constructors

TypeModel

Fields

typeName :: Type exRef

typeEnv :: TypeEnv adtName consName inRef exRef

data TypeN r

Constructors

TypeN r [TypeN r]

data TypeRef name

Constructors

TypVar Word8

TypRef name

type Errors = [Error]

type family Ana t where ...

Equations

Ana (f a0 a1 a2 a3 a4 a5 a6 a7 a8) = App (App (App (App (App (App (App (App (App (Typ (f A0 A1 A2 A3 A4 A5 A6 A7 A8)) (Ana a0)) (Ana a1)) (Ana a2)) (Ana a3)) (Ana a4)) (Ana a5)) (Ana a6)) (Ana a7)) (Ana a8)

Ana (f a0 a1 a2 a3 a4 a5 a6 a7) = App (App (App (App (App (App (App (App (Typ (f A0 A1 A2 A3 A4 A5 A6 A7)) (Ana a0)) (Ana a1)) (Ana a2)) (Ana a3)) (Ana a4)) (Ana a5)) (Ana a6)) (Ana a7)

Ana (f a0 a1 a2 a3 a4 a5 a6) = App (App (App (App (App (App (App (Typ (f A0 A1 A2 A3 A4 A5 A6)) (Ana a0)) (Ana a1)) (Ana a2)) (Ana a3)) (Ana a4)) (Ana a5)) (Ana a6)

Ana (f a0 a1 a2 a3 a4 a5) = App (App (App (App (App (App (Typ (f A0 A1 A2 A3 A4 A5)) (Ana a0)) (Ana a1)) (Ana a2)) (Ana a3)) (Ana a4)) (Ana a5)

Ana (f a0 a1 a2 a3 a4) = App (App (App (App (App (Typ (f A0 A1 A2 A3 A4)) (Ana a0)) (Ana a1)) (Ana a2)) (Ana a3)) (Ana a4)

Ana (f a0 a1 a2 a3) = App (App (App (App (Typ (f A0 A1 A2 A3)) (Ana a0)) (Ana a1)) (Ana a2)) (Ana a3)

Ana (f a0 a1 a2) = App (App (App (Typ (f A0 A1 A2)) (Ana a0)) (Ana a1)) (Ana a2)

Ana (f a0 a1) = App (App (Typ (f A0 A1)) (Ana a0)) (Ana a1)

Ana (f a0) = App (Typ (f A0)) (Ana a0)

Ana a = Typ a

module ZM.Abs

data TypedValue a

A typed value, a value and its absolute type

Constructors

TypedValue AbsType a

data TypedBLOB

A typed value, a Flat encoded value and its absolute type

Constructors

TypedBLOB AbsType (BLOB FlatEncoding)

data BLOB encoding

A BLOB is binary value encoded according to a specified encoding (e.g. UTF8)

Constructors

BLOB encoding ByteString

unblob :: BLOB t -> ByteString

Extract the binary content of a BLOB

blob :: AsByteString a => encoding -> a -> BLOB encoding

Build a BLOB from an encoding and a ByteString-like value

typedBLOB :: forall a. (Model a, Flat a) => a -> TypedBLOB

Build a TypedBLOB out of a value

typedBLOB_ :: Flat a => AbsType -> a -> TypedBLOB

Build a TypedBLOB out of a type and a value

typedValue :: forall a. Model a => a -> TypedValue a

Build a TypedValue out of a value

untypedBLOB :: forall a. (Flat a, Model a) => Decoded TypedBLOB -> TypedDecoded a

Type-checked extraction of a value of a known type from a decoded TypedBLOB

untypedValue :: Model a => Decoded (TypedValue a) -> TypedDecoded a

Type-checked extraction of a value of a known type from a decoded TypedValue

typeErr :: AbsType -> AbsType -> TypedDecoded a

Return a WrongType error

module ZM.Pretty

module ZM.Transform

module ZM.Types