ZM.Abs

Description

Derive absolute/canonical data type models

Documentation

absEnv :: Model a => Proxy a -> AbsEnv

absType :: Model a => Proxy a -> AbsType

Derive an absolute type for a type, or throw an error if derivation is impossible

absTypeModel :: Model a => Proxy a -> AbsTypeModel

Derive an absolute type model for a type, or throw an error if derivation is impossible

absTypeModelMaybe :: Model a => Proxy a -> Either [ZMError QualName] AbsTypeModel

Derive an absolute type model for a type, composed of the type plus its full type environment (all the types the type depends upon directly or indirectly)

>>> let Right t = absTypeModelMaybe (Proxy :: Proxy [Bool]) in error . prettyShow $ t
Type: List Bool (Kb8cd13187198 K306f1981b41c)
Environment:
Bool.K306f1981b41c ≡   False
                     | True;
List.Kb8cd13187198 a ≡   Nil
                       | Cons a (List.Kb8cd13187198 a);

Mutual dependencies are not supported:

>>> data A = A B deriving (Model,Generic);data B = B A deriving (Model,Generic)
>>> error . prettyShow $ absTypeModelMaybe (Proxy :: Proxy A)
Left [Found mutually recursive types: [interactive.Ghci2.B,
                                       interactive.Ghci2.A]]

Higher kind types are also not supported (this is detected at compile time as it is not possible to derive a Model instance):

>>> data Free f a = Impure (f (Free f a)) | Pure a deriving Model
Could not deduce (AsType (Ana a))
  arising from the 'deriving' clause of a data type declaration
from the context: (Typeable f, Typeable a)
  bound by the deriving clause for ‘Model (Free f a)’
  at <interactive>:1:58-62
Possible fix:
  use a standalone 'deriving instance' declaration,
    so you can specify the instance context yourself

relToAbsEnv :: (IsKey r, IsId n) => RelTypeEnv n r -> Either [ZMError r] AbsEnv

Convert a set of relative ADTs to the equivalent ZM absolute ADTs

relToAbsEnvWith :: (IsKey r, IsId n) => AbsEnv -> RelTypeEnv n r -> Either [ZMError r] AbsEnv

Convert a set of relative ADTs to the equivalent ZM absolute ADTs, with a starting set of known absolute ADTs.

Conversion will fail if the relative ADTs are mutually recursive or refer to undefined ADTs.

refErrors :: (Foldable t, Eq t1) => Map t1 (t (ADTRef t1)) -> [ZMError t1]

Check that all internal references in the ADT definitions are to ADTs defined in the environment This test is also performed as part of the relative to absolute conversion (relToAbsEnv)

kindErrors :: Ord k => Map k (ADT name1 name2 (TypeRef2 a k)) -> [ZMError (Either a k)]

data TypeRef2 v r

Check that all type expressions have kind *, that's to say:

Type constructors are fully applied

Type variables have * kind

Constructors

Var2 v

Ext2 r

class KeyOf k a where

Methods

keyOf :: a -> k

transitiveClosure :: (Foldable t, Ord r) => (a -> Maybe r) -> Map r (t a) -> r -> Either [ZMError r] [r]

Return the transitive closure of an element in a graph of dependencies specified as an adjacency list

>>> transitiveClosure Just (M.fromList [("a",["b","c"]),("b",["b","d","d","c"]),("c",[]),("d",["a"])]) "b"
Right ["c","a","d","b"]

>>> transitiveClosure Just (M.fromList [("a",["b","c"]),("b",["b","d","d","c"]),("c",[]),("d",["a"])]) "c"
Right ["c"]