ZM.BLOB

Documentation

data BLOB encoding

A BLOB is binary value encoded according to a specified encoding (e.g. UTF8)

Constructors

BLOB

Fields

encoding :: encoding

content :: ByteString

blob :: AsByteString a => encoding -> a -> BLOB encoding

Build a BLOB from an encoding and a ByteString-like value

unblob :: BLOB t -> ByteString

Extract the binary content of a BLOB

data TypedBLOB

A typed value, a Flat encoded value and its absolute type

Constructors

TypedBLOB AbsType (BLOB FlatEncoding)

typedBLOB :: forall a. (Model a, Flat a) => a -> TypedBLOB

Build a TypedBLOB out of a value

typedBLOB_ :: Flat a => AbsType -> a -> TypedBLOB

Build a TypedBLOB out of a type and a value

untypedBLOB :: forall a. (Flat a, Model a) => Decoded TypedBLOB -> TypedDecoded a

Type-checked extraction of a value of a known type from a decoded TypedBLOB

data TypedValue a

A typed value, a value and its absolute type

Constructors

TypedValue AbsType a

typedValue :: forall a. Model a => a -> TypedValue a

Build a TypedValue out of a value

untypedValue :: Model a => Decoded (TypedValue a) -> TypedDecoded a

Type-checked extraction of a value of a known type from a decoded TypedValue

typeErr :: AbsType -> AbsType -> TypedDecoded a

Return a WrongType error