ZM.Parser.ADT

Documentation

adts :: Parser [ADTParts]

Parse a, possibly empty, group of ZM ADT declarations.

>>> parseMaybe adts ""
Just []

>>> putStr $ prettyShow $ parseMaybe adts "Void;"
Just [Void@(0:0-3) =]

>>> putStr $ prettyShow $ parseMaybe adts "Void;Bool=False | True"
Just [Void@(0:0-3) =,
      Bool@(0:5-8) = False@(0:10-14) | True@(0:18-21)]

adt :: Parser ADTParts

Parse a ZM ADT declaration.

>>> putStr $ prettyShow $ parseMaybe adt "Void"
Just Void@(0:0-3) =

>>> putStr $ prettyShow $ parseMaybe adt "Bool.K306f1981b41c = False \n| True"
Just Bool.K306f1981b41c@(0:0-17) = False@(0:21-25) | True@(1:2-5)

>>> putStr $ prettyShow $ parseMaybe adt "是不是 ≡   是\n                    | 不是"
Just 是不是@(0:0-2) = 是@(0:8) | 不是@(1:22-23)

>>> putStr $ prettyShow $ parseMaybe adt "A = A (A (B C D))"
Just A@(0:0) = A@(0:4) (A@(0:7) (B@(0:10) C@(0:12) D@(0:14)))

>>> putStr $ prettyShow $ parseMaybe adt "Msg = Msg {from:Name.K306f1981b41c,subject:String}"
Just Msg@(0:0-2) = Msg@(0:6-8) {from@(0:11-14) :: Name.K306f1981b41c@(0:16-33),
                                subject@(0:35-41) :: String@(0:43-48)}

>>> putStr $ prettyShow $ parseMaybe adt "List a = Cons {head:a, tail:List a} | \nNil"
Just List@(0:0-3) a@(0:5) = Cons@(0:9-12) {head@(0:15-18) :: a@(0:20),
                                           tail@(0:23-26) :: List@(0:28-31) a@(0:33)} | Nil@(1:0-2)

Syntax Errors:

>>> putStr $ prettyShow $ parseDoc adt "Bool = | True"
Left "unexpected '|' expecting end of input or letter"@(0:7)

>>> putStr $ prettyShow $ parseDoc adt " = "
Left "unexpected '=' expecting letter"@(0:1)

>>> putStr $ prettyShow $ parseDoc adt "a b.d e.f = c.f.a"
Left "unexpected '.' expecting '=', '_', '\8801', alphanumeric character, end of input, or letter"@(0:3)

parType :: Parser a -> Parser (Type a)

Parse a type application, a type constructor with zero or more parameters

>>> parseMaybe (parType localId) "a"
Just (TypeCon "a")

>>> parseMaybe (parType localId) "((a))"
Just (TypeCon "a")

>>> parseMaybe (parType localId) "a b"
Just (TypeApp (TypeCon "a") (TypeCon "b"))

>>> parseMaybe (parType localId) "(a) ((b)) c"
Just (TypeApp (TypeApp (TypeCon "a") (TypeCon "b")) (TypeCon "c"))

>>> typeN <$> parseMaybe (parType localId) "(a b (c d (e f g))) z"
Just (TypeN "a" [TypeN "b" [],TypeN "c" [TypeN "d" [],TypeN "e" [TypeN "f" [],TypeN "g" []]],TypeN "z" []])

>>> parseMaybe (parType localId) "(a b) c"
Just (TypeApp (TypeApp (TypeCon "a") (TypeCon "b")) (TypeCon "c"))

absReference :: Parser AbsRef

>>> prettyShow <$> parseMaybe absReference "K306f1981b41c"
Just "K306f1981b41c"

>>> prettyShow <$> parseMaybe absReference "KKK"
Nothing

maybeNamedAbsRef :: Parser (TypeName Identifier)

>>> prettyShow <$> parseMaybe maybeNamedAbsRef "Bool.K306f1981b41c"
Just "Bool.K306f1981b41c"

>>> prettyShow <$> parseMaybe maybeNamedAbsRef ".K306f1981b41c"
Just ".K306f1981b41c"

>>> prettyShow <$> parseMaybe maybeNamedAbsRef "K306f1981b41c"
Just ".K306f1981b41c"

>>> prettyShow <$> parseMaybe maybeNamedAbsRef "Bool"
Nothing

namedOrAbsRef :: Parser (TypeName Identifier)

Parse a simple type name, possibly qualified with an absolute code.

>>> prettyShow <$> parseMaybe namedOrAbsRef "nil"
Just "nil"

The . is required to distinguish a hash from a plain type name (remove this ambiguity?)

This is a hash:

>>> parseMaybe namedOrAbsRef ".K306f1981b41c"
Just (That (AbsRef (SHAKE128_48 48 111 25 129 180 28)))

And this is just a type name (NO actually we try to interpret it as an absolute reference first):

>>> parseMaybe namedOrAbsRef "K306f1981b41c"
Just (That (AbsRef (SHAKE128_48 48 111 25 129 180 28)))