ZM.Parser.Doc

Documentation

>>> import Text.Megaparsec(parseMaybe)

data Section

Constructors

Section

Fields

name :: String

body :: String

sectionText :: Section -> String

sectionsText :: [Section] -> String

data P b

Constructors

forall a. P

Fields

parser :: Parser a

converter :: a -> b

evalSections :: Map String (Parser a) -> String -> Either String [a]

Document evaluation

Variation: Section parsers, in addition of their content are also given access to: * their context (the doc and their position in it) * the parser itself so thaty can implement nested parsing. * a read/write environment [(String,a)]

{|This is some text with an embedded graph {dot ...} }

>>> data R = AChar Char | AStr String deriving (Show)
>>> parsers = M.fromList [("char", AChar <$> charLiteral), ("$", AStr <$> stringLiteral)]

>>> parsers = M.fromList [("char", Left <$> charLiteral), ("$", Right <$> stringLiteral)]
>>> evalSections parsers "char 'c'\n$ \"abc\""

parseS :: Parser a -> String -> Either String a

sections :: Parser [Section]

>>> parseMaybe sections "-- a comment\nadt Bool = False \n | True\nadt Enum = A \n   | B\n | C"

parseMaybe sections "adt Bool = False n | Truenadt Enum = A n | Bn | C"

>>> let src = "adt Bool = False \n | True\nadt Enum = A \n   | B\n | C" in (\ss -> src == sectionsText ss) <$> (parseMaybe sectionsDoc src)

>>> parseMaybe sectionsDoc "\n  abc \nadt Bool = False \n| True\nadt Enum = A \n   | B\n | C"
Nothing

>>> parseMaybe sectionsDoc "\n \t \r\n \nadt Bool = False \n| True\nadt Enum = A \n   | B\n | C"

sectionsDoc :: Parser [Section]

>>> let src = "adt Bool = False \n | True\nadt Enum = A \n   | B\n | C" in (\ss -> src == sectionsText ss) <$> (parseMaybe sectionsDoc src)

>>> parseMaybe sectionsDoc "\n  abc \nadt Bool = False \n| True\nadt Enum = A \n   | B\n | C"
Nothing

>>> parseMaybe sectionsDoc "\n \t \r\n \nadt Bool = False \n| True\nadt Enum = A \n   | B\n | C"

section :: Parser Section

>>> parseMaybe section "adt"
Just (Section {name = "adt", body = ""})

>>> parseMaybe section "adt    Bool = False | True"
Just (Section {name = "adt", body = "Bool = False | True"})

>>> parseMaybe section "adt Enum = A \n   | B\n | C"
Just (Section {name = "adt", body = "Enum = A \n   | B\n | C"})

sectionBody :: Parser String

sectionEnd :: Parser ()

elems :: ParsecT Void String Identity [Either String [Char]]

anyChar :: Parser Char

sectionName :: Parser String

>>> parseMaybe sectionName " adt"
Nothing

>>> parseMaybe sectionName "adt"
Just "adt"

>>> parseMaybe sectionName "-()[]{}"
Nothing

>>> parseMaybe sectionName ":+*/.@#$%?><;"
Just ":+*/.@#$%?><;"

>>> parseMaybe sectionName "_"

>>> parseMaybe sectionName "{"

isNameChar :: Char -> Bool

skipEmptyLines :: Parser ()

>>> parseMaybe skipEmptyLines ""

>>> parseMaybe skipEmptyLines "\n   \t  \n"
Just ()

>>> parseMaybe skipEmptyLines "\n "
Nothing

skipEmptyLine :: Parser ()