ZM.Parser.Types

Documentation

type Parser = Parsec Void String

data ADTParts

A parsed ADT

Constructors

ADTParts

Fields

name :: AtAbsName

vars :: [AtId]

constrs :: [(AtId, Fields AtId AtAbsName)]

asTypeName :: Maybe a -> Maybe b -> These a b

hasName :: These a b -> Bool

hasRef :: These a1 a -> Bool

type TypeName l = These l AbsRef

A data type name can be either local, or absolute, or both: Bool | ".K306f1981b41c" | Bool.K306f1981b41c

localTypeName :: These a b -> a

absTypeName :: These a b -> b

data Label l a

Constructors

Label

Fields

label :: l

object :: a

type At v = Label Range v

type AtId = At Identifier

type AtAbsName = At (TypeName Identifier)

type AtError = At String

data Range

Constructors

Range

Fields

line, start, end :: Word32

type Val lit binder = Fix (ValF lit binder)

data ValF lit binder r

Constructors

ConstrF

Fields

String

Name of the constructor (e.g. True)

(Either [r] [(String, r)])

Constructor parameters, possibly named } -- ^A standard constructor, e.g. PCon True []

LitF (lit r)
A variable or a lit pattern (e.g. a string or a number)

BinderF binder
A pattern that might bind a matched value to a name

pattern Constr :: String -> Either [Val lit binder] [(String, Val lit binder)] -> Val lit binder

pattern VInteger :: Integer -> Val Literal binder

pattern VChar :: Char -> Val Literal binder

pattern VString :: String -> Val Literal binder

pattern VFloat :: Double -> Val Literal binder

pattern VArray :: [Val Literal binder] -> Val Literal binder

pattern PBind :: String -> Val lit Binder

pattern PWild :: Val lit Binder

newtype Fix (f :: Type -> Type)

Constructors

Fix

Fields

unFix :: f (Fix f)

data Annotate label f

Constructors

Annotate label (f (Annotate label f))

data Literal e

Constructors

LInteger Integer

LFloat Double

LChar Char

LString String

LArray [e]

data Binder

Constructors

Wild

Bind String

type Value = Annotate (AbsType, [Bool]) (ValF Literal Void)

pattern Value :: forall a b lit binder. a -> String -> b -> [Annotate (a, b) (ValF lit binder)] -> Annotate (a, b) (ValF lit binder)

valType :: forall a b (lit :: Type -> Type) binder. Annotate (a, b) (ValF lit binder) -> a

valName :: forall a b (lit :: Type -> Type) binder. Annotate (a, b) (ValF lit binder) -> String

valBits :: forall a b (lit :: Type -> Type) binder. Annotate (a, b) (ValF lit binder) -> b

valFields :: forall a b (lit :: Type -> Type) binder. Annotate (a, b) (ValF lit binder) -> [Annotate (a, b) (ValF lit binder)]

data Void

Uninhabited data type

Since: base-4.8.0.0

data Void1 a

A generic value (used for dynamic decoding) data Value = Value { valType :: AbsType -- ^Type , valName :: String -- ^Constructor name (duplicate info if we have abstype) , valBits :: [Bool] -- ^Bit encoding/constructor id -- TODO: add field names (same info present in abstype) , valFields :: [Value] -- ^Values to which the constructor is applied, if any } deriving (Eq, Ord, Show, NFData, Generic, Flat) data Void Value could also be expressed as type Value = Val (AbsType, [Bool]) Literal Void v0 :: [Value0] v0 = [lit (.., [False,True]) (PInt 3) |A Pattern matches a subset of values of a ZM Type p1 :: [Pattern] p1 = [lit "1:1" (PInt 3), Binder "1:2" PWild] type Pattern = Val () Literal Binder type Value = Val () Literal () data Val annotation lit binder = Constr annotation String -- ^Name of the constructor (e.g. True) (Either [Val annotation lit binder] [(String, Val annotation lit binder)]) -- ^Constructor parameters, possibly named -- } -- ^A standard constructor, e.g. PCon True []

Orphan instances

Pretty n => Pretty (TypeName n)

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> TypeName n -> Doc

pPrint :: TypeName n -> Doc

pPrintList :: PrettyLevel -> [TypeName n] -> Doc