ZM.Pretty

Description

Pretty instances for some basic Haskell types and for data type models

Documentation

hex :: Word8 -> String

Display a Word in hexadecimal format

unPrettyRef :: String -> SHAKE128_48 a

Convert the textual representation of a hash code to its equivalent value

>>> unPrettyRef "Kb53bec846608"
SHAKE128_48 181 59 236 132 102 8

>>> unPrettyRef "Kb53bec8466080000"
*** Exception: unPrettyRef: unknown code "b53bec8466080000"
...

prettyList :: [Doc] -> Doc

Display a list of Docs, with spaced elements >>> prettyList (map pPrint [11,22,33::Word8]) [11, 22, 33]

prettyTuple :: [Doc] -> Doc

Display a list of Docs, as a tuple with spaced elements

>>> prettyTuple (map pPrint [11,22,33::Word8])
(11, 22, 33)

Orphan instances

Pretty Int8

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> Int8 -> Doc

pPrint :: Int8 -> Doc

pPrintList :: PrettyLevel -> [Int8] -> Doc

Pretty Int16

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> Int16 -> Doc

pPrint :: Int16 -> Doc

pPrintList :: PrettyLevel -> [Int16] -> Doc

Pretty Int32

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> Int32 -> Doc

pPrint :: Int32 -> Doc

pPrintList :: PrettyLevel -> [Int32] -> Doc

Pretty Int64

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> Int64 -> Doc

pPrint :: Int64 -> Doc

pPrintList :: PrettyLevel -> [Int64] -> Doc

Pretty Word

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> Word -> Doc

pPrint :: Word -> Doc

pPrintList :: PrettyLevel -> [Word] -> Doc

Pretty Word8

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> Word8 -> Doc

pPrint :: Word8 -> Doc

pPrintList :: PrettyLevel -> [Word8] -> Doc

Pretty Word16

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> Word16 -> Doc

pPrint :: Word16 -> Doc

pPrintList :: PrettyLevel -> [Word16] -> Doc

Pretty Word32

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> Word32 -> Doc

pPrint :: Word32 -> Doc

pPrintList :: PrettyLevel -> [Word32] -> Doc

Pretty Word64

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> Word64 -> Doc

pPrint :: Word64 -> Doc

pPrintList :: PrettyLevel -> [Word64] -> Doc

Pretty ShortByteString

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> ShortByteString -> Doc

pPrint :: ShortByteString -> Doc

pPrintList :: PrettyLevel -> [ShortByteString] -> Doc

Pretty ByteString

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> ByteString -> Doc

pPrint :: ByteString -> Doc

pPrintList :: PrettyLevel -> [ByteString] -> Doc

Pretty ByteString

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> ByteString -> Doc

pPrint :: ByteString -> Doc

pPrintList :: PrettyLevel -> [ByteString] -> Doc

Pretty Text

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> Text -> Doc

pPrint :: Text -> Doc

pPrintList :: PrettyLevel -> [Text] -> Doc

Pretty UTF16Text

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> UTF16Text -> Doc

pPrint :: UTF16Text -> Doc

pPrintList :: PrettyLevel -> [UTF16Text] -> Doc

Pretty UTF8Text

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> UTF8Text -> Doc

pPrint :: UTF8Text -> Doc

pPrintList :: PrettyLevel -> [UTF8Text] -> Doc

Pretty NoEncoding

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> NoEncoding -> Doc

pPrint :: NoEncoding -> Doc

pPrintList :: PrettyLevel -> [NoEncoding] -> Doc

Pretty TypedDecodeException

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> TypedDecodeException -> Doc

pPrint :: TypedDecodeException -> Doc

pPrintList :: PrettyLevel -> [TypedDecodeException] -> Doc

Pretty Identifier
Convert references in an absolute definition to their textual form (useful for display) stringADT :: AbsEnv -> AbsADT -> ADT Identifier Identifier (TypeRef Identifier) stringADT env adt = let name = declName adt in ADT name (declNumParameters adt) ((solveS name $) $ declCons adt) where solveS _ (Var n) = TypVar n solveS _ (Ext k) = TypRef . declName . solve k $ env solveS name Rec = TypRef name

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> Identifier -> Doc

pPrint :: Identifier -> Doc

pPrintList :: PrettyLevel -> [Identifier] -> Doc

Pretty AbsEnv

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> AbsEnv -> Doc

pPrint :: AbsEnv -> Doc

pPrintList :: PrettyLevel -> [AbsEnv] -> Doc

Pretty AbsTypeModel

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> AbsTypeModel -> Doc

pPrint :: AbsTypeModel -> Doc

pPrintList :: PrettyLevel -> [AbsTypeModel] -> Doc

Pretty a => Pretty (Seq a)

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> Seq a -> Doc

pPrint :: Seq a -> Doc

pPrintList :: PrettyLevel -> [Seq a] -> Doc

Pretty a => Pretty (NonEmptyList a)

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> NonEmptyList a -> Doc

pPrint :: NonEmptyList a -> Doc

pPrintList :: PrettyLevel -> [NonEmptyList a] -> Doc

Pretty a => Pretty (ADTRef a)

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> ADTRef a -> Doc

pPrint :: ADTRef a -> Doc

pPrintList :: PrettyLevel -> [ADTRef a] -> Doc

Pretty a => Pretty (String, ADTRef a)

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> (String, ADTRef a) -> Doc

pPrint :: (String, ADTRef a) -> Doc

pPrintList :: PrettyLevel -> [(String, ADTRef a)] -> Doc

Pretty (AbsEnv, AbsADT)

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> (AbsEnv, AbsADT) -> Doc

pPrint :: (AbsEnv, AbsADT) -> Doc

pPrintList :: PrettyLevel -> [(AbsEnv, AbsADT)] -> Doc

Pretty (AbsEnv, AbsType)

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> (AbsEnv, AbsType) -> Doc

pPrint :: (AbsEnv, AbsType) -> Doc

pPrintList :: PrettyLevel -> [(AbsEnv, AbsType)] -> Doc

(Pretty a, Pretty b) => Pretty (Map a b)

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> Map a b -> Doc

pPrint :: Map a b -> Doc

pPrintList :: PrettyLevel -> [Map a b] -> Doc

(Pretty a, Pretty b, Pretty c, Pretty d, Pretty e, Pretty f, Pretty g, Pretty h, Pretty i) => Pretty (a, b, c, d, e, f, g, h, i)

Instance details

Methods

pPrintPrec :: PrettyLevel -> Rational -> (a, b, c, d, e, f, g, h, i) -> Doc

pPrint :: (a, b, c, d, e, f, g, h, i) -> Doc

pPrintList :: PrettyLevel -> [(a, b, c, d, e, f, g, h, i)] -> Doc