ZM.To.Encoder

Documentation

encoderFor :: Model a => Proxy a -> ParserEncoder

Return a Flat encoder for values of the ZM type corresponding to the provided Haskell type

typeEncoder :: AbsTypeModel -> ParserEncoder

Return a Flat encoder for values of the provided ZM type