ZM.Transform

Description

Utilities to operate on the absolute type model

Saturated ADTs

type MapTypeTree = Map (Type AbsRef) (ConTree Identifier AbsRef)

A map of fully applied types to the corresponding saturated constructor tree

typeTree :: AbsTypeModel -> MapTypeTree

Return the map of types to saturated constructor trees corresponding to the type model

solvedADT :: (Ord ref, Show ref) => Map ref (ADT name consName (ADTRef ref)) -> Type ref -> ADT name consName ref

Convert a type to an equivalent concrete ADT whose variables have been substituted by the type parameters (e.g. Maybe Bool -> Maybe = Nothing | Just Bool)

Dependencies

typeDefinition :: AbsEnv -> AbsType -> Either [ZMError AbsRef] [AbsADT]

Return all the ADTs referred, directly or indirectly, by the provided type, and defined in the provided environment

adtDefinition :: AbsEnv -> AbsRef -> Either [ZMError AbsRef] [AbsADT]

Return all the ADTs referred, directly or indirectly, by the ADT identified by the provided reference, and defined in the provided environment

innerReferences :: AbsADT -> [AbsRef]

Return the list of references found in the ADT definition

references :: AbsType -> [AbsRef]

Return the list of references found in the absolute type