ZM.Type.Array

Documentation

data Array a

An Array.

A sequence of sequences of up to to 255 values. Array a = A0 | A1 a (Array a) | A2 a a (Array a) ... | A255 a ... (Array a)

Constructors

Array [a]

data Bytes

A byte-aligned byte array To encode and decode efficiently an Array of Bytes, we pre-align it to the nearest byte border.

Constructors

Bytes (PreAligned (Array Word8))

Orphan instances

Model Filler

Instance details

Methods

envType :: Proxy Filler -> State Env HType

Model a => Model (PreAligned a)

Instance details

Methods

envType :: Proxy (PreAligned a) -> State Env HType