ZM.Type.Words

Documentation

data Sign

Constructors

Positive

Negative

data Word7

A 7 bits unsigned integer data Word7 = V0 .. V127

Constructors

Word7 Word8

data Word

An unsigned integer of arbitrary length encoded as a non empty list of Word7 words with least significant word first and, inside each word, most significant bit first.

Example: 3450 :: Word

Binary representation: 0000110101111010

Split in 7bits groups: 0011010(==26 decimal) 1111010(==122 decimal)

Build a non-empty list whose elements are the groups in reverse order: Word (Cons V122 (Elem V26))

So Least Significant Byte first with Most Significant Bit first in every 7 bits group.

- BUG: MostSignificantFirst is useless, Word7 already has an order.

Constructors

Word (LeastSignificantFirst (NonEmptyList (MostSignificantFirst Word7)))

data Word8

An 8 bits unsigned integer data Word8 = V0 | V1 .. | V255

Constructors

Word8 Word8

data Word16

Constructors

Word16 Word

data Word32

Constructors

Word32 Word

data Word64

Constructors

Word64 Word

data Int

Constructors

Int (ZigZag Word)

data Int8

Constructors

Int8 (ZigZag Word8)

data Int16

Constructors

Int16 (ZigZag Word16)

data Int32

Constructors

Int32 (ZigZag Word32)

data Int64

Constructors

Int64 (ZigZag Word64)

data ZigZag a

ZigZag encoding, map signed integers to unsigned integers Positive integers are mapped to even unsigned values, negative integers to odd values: 0 -> 0, -1 -> 1, 1 -> 2, -2 -> 3, 2 -> 4 ...

Constructors

ZigZag a

data MostSignificantFirst a

Constructors

MostSignificantFirst a

data LeastSignificantFirst a

Constructors

LeastSignificantFirst a