ZM.Util

Documentation

proxyOf :: a -> Proxy a

Return the proxy for the type of the given value

State Monad utilities

runEnv :: State (Map k a1) a -> (a, Map k a1)

Run a State monad with an empty map as environment

execEnv :: State (Map k a1) a -> Map k a1

Exec a State monad with an empty map as environment